server: Add `OLLAMA_PULL_TIMEOUT` environment variable to reduce stall occurrences
In poor network conditions, such as when QoS is applied, downloading with the current 16 parts often results in only one part being transmitted while the remaining 15 parts are not. This leads to a stalled message after 30 seconds of inactivity, causing the progress to regress. The goal is to reduce stall occurrences by allowing the adjustment of timeout duration of parts.

In cases of poor network conditions, it is recommended to set `OLLAMA_PULL_TIMEOUT` to over than 6000.